reporttask package refactored
