+++ src/main/kotlin/io/github/janbarari/gradle/analytics/reporttask/render/InitialRenderStage.kt
@@ -1,4 +1,4 @@
-package io.github.janbarari.gradle.analytics.reporttask
+package io.github.janbarari.gradle.analytics.reporttask.render
 
 import io.github.janbarari.gradle.analytics.GradleAnalyticsPlugin
 import io.github.janbarari.gradle.analytics.core.Stage
